perf: separate layout-dirty from visual-dirty tracking and optimize hit testing
- Separate needs_layout (taffy recomputation) from needs_sync (scene-only
  sync) so visual-only changes like opacity/translate animations skip the
  expensive taffy compute_layout pass entirely
- Add Style::layout_eq/visual_eq and WidgetProps::layout_eq for granular
  change detection in the reconciler — only call taffy.set_style when
  layout-affecting properties actually changed
- Cache absolute bounds on UiNode during sync so hit_test uses O(1) cached
  lookups instead of per-call taffy.layout() tree walks
- Change dirty_components from Vec to HashSet for O(1) insert/contains
- Use swap_remove instead of remove for O(h) → O(1) template handler
  extraction in resolve.rs
- Fix clippy warnings: unnecessary i32 casts in video player, collapsible_if,
  std::io::Error::other

+++ crates/yumeri-video/src/decode/vulkan.rs
@@ -236,7 +236,7 @@ impl DecoderBackend for VulkanDecoder {
         };
 
         let format = unsafe { (*frame.as_ptr()).format };
-        if format != ffi::AV_PIX_FMT_VULKAN as i32 {
+        if format != ffi::AV_PIX_FMT_VULKAN {
             // FFmpeg didn't produce Vulkan frames — fall back to CPU conversion
             return self.decode_sw_fallback(&frame);
         }
